Moved members package into list to keep resource/subresource/subsubresource hierarchy
