+++ src/main/java/com/ecwid/maleorang/method/v3_0/batches/DeleteBatchMethod.kt
@@ -1,12 +1,11 @@
-package com.ecwid.maleorang.method.v3_0.members
+package com.ecwid.maleorang.method.v3_0.lists.members
 
-import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.MailchimpMethod
+import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.APIVersion
 import com.ecwid.maleorang.annotation.HttpMethod
 import com.ecwid.maleorang.annotation.Method
 import com.ecwid.maleorang.annotation.PathParam
-import org.apache.commons.codec.digest.DigestUtils
 
 /**
  * [Remove a list member](http://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/#delete-delete_lists_list_id_members_subscriber_hash)
